+++ src/fonduer/parser/preprocessors/__init__.py
@@ -1,11 +1,11 @@
-from fonduer.parser.preprocessors.csv_paths_preprocessor import CSVPathsPreprocessor
+from fonduer.parser.preprocessors.csv_doc_preprocessor import CSVDocPreprocessor
 from fonduer.parser.preprocessors.doc_preprocessor import DocPreprocessor
 from fonduer.parser.preprocessors.html_doc_preprocessor import HTMLDocPreprocessor
 from fonduer.parser.preprocessors.text_doc_preprocessor import TextDocPreprocessor
 from fonduer.parser.preprocessors.tsv_doc_preprocessor import TSVDocPreprocessor
 
 __all__ = [
-    "CSVPathsPreprocessor",
+    "CSVDocPreprocessor",
     "DocPreprocessor",
     "HTMLDocPreprocessor",
     "TSVDocPreprocessor",